Add guard to `purge` command (#115)

+++ blockservice/src/cmd/purge.rs
@@ -1,4 +1,4 @@
-use std::path::Path;
+use std::{io::Write, path::Path};
 
 use crate::{app_dir::open_app_dir, db::BlockDb};
 
@@ -7,28 +7,41 @@ pub fn purge(
     chain_id: u64,
     from: Option<u64>,
     to: Option<u64>,
+    mut reader: impl std::io::BufRead,
 ) -> Result<(), Box<dyn std::error::Error>> {
-    let db = open_app_dir(app_dir, false)?;
-
-    db.delete_range(chain_id, from, to)?;
-
+    // Guard the purge command
+    let mut input = String::new();
+    print!("Are you sure you want to purge blocks for chain {chain_id}? (y/n): ");
+    std::io::stdout().flush()?;
+    reader.read_line(&mut input)?;
+    if matches!(input.trim(), "y" | "Y") {
+        let db = open_app_dir(app_dir, false)?;
+        db.delete_range(chain_id, from, to)?;
+    }
     Ok(())
 }
 
 #[cfg(test)]
 mod tests {
-    use std::os::unix::fs::PermissionsExt;
+    use std::{io::Cursor, os::unix::fs::PermissionsExt, vec};
 
     use bertha_types::Block;
 
     use super::*;
     use crate::app_dir::{BLOCK_DB_NAME, init_app_dir};
 
+    /// Helper function to simulate user confirmation for the purge command.
+    fn confirm_purge(bool: bool) -> impl std::io::BufRead {
+        let input = if bool { "y\n" } else { "n\n" };
+        Cursor::new(input)
+    }
+
     #[test]
     fn fails_if_app_dir_is_not_initialized() {
         let tmpdir = tempfile::tempdir().unwrap();
 
-        let result = purge(tmpdir.path(), 0, None, None);
+        // set the stdin
+        let result = purge(tmpdir.path(), 0, None, None, confirm_purge(true));
         assert!(result.is_err());
         assert!(result.unwrap_err().to_string().contains(&format!(
             "no database found at {} - did you forget to run init?",
@@ -50,7 +63,7 @@ mod tests {
         )
         .unwrap();
 
-        let result = purge(tmpdir.path(), 0, None, None);
+        let result = purge(tmpdir.path(), 0, None, None, confirm_purge(true));
         assert!(result.is_err());
         assert!(
             result
@@ -66,9 +79,9 @@ mod tests {
 
         init_app_dir(tmpdir.path()).unwrap();
 
-        assert!(purge(tmpdir.path(), 0, None, None).is_ok());
-        assert!(purge(tmpdir.path(), 0, Some(0), None).is_ok());
-        assert!(purge(tmpdir.path(), 0, Some(0), Some(1)).is_ok());
+        assert!(purge(tmpdir.path(), 0, None, None, confirm_purge(true)).is_ok());
+        assert!(purge(tmpdir.path(), 0, Some(0), None, confirm_purge(true)).is_ok());
+        assert!(purge(tmpdir.path(), 0, Some(0), Some(1), confirm_purge(true)).is_ok());
     }
 
     #[test]
@@ -93,12 +106,66 @@ mod tests {
 
         drop(db); // close the database to ensure that the purge command can open it
 
-        purge(tmpdir.path(), chain_id, Some(1), Some(2)).unwrap();
+        purge(
+            tmpdir.path(),
+            chain_id,
+            Some(1),
+            Some(2),
+            confirm_purge(true),
+        )
+        .unwrap();
 
         let db = open_app_dir(tmpdir.path(), false).unwrap();
         assert!(db.get(chain_id, 0).unwrap().is_some());
         assert!(db.get(chain_id, 1).unwrap().is_none());
         assert!(db.get(chain_id, 2).unwrap().is_none());
         assert!(db.get(chain_id, 3).unwrap().is_some());
+    }
+
+    #[test]
+    fn cancel_operation_if_user_does_not_confirm() {
+        let tmpdir = tempfile::tempdir().unwrap();
+        init_app_dir(tmpdir.path()).unwrap();
+
+        let db = open_app_dir(tmpdir.path(), false).unwrap();
+        db.put_raw(42, 1, vec![1, 2, 3].as_slice()).unwrap();
+
+        purge(tmpdir.path(), 0, None, None, confirm_purge(false)).expect("purge should succeed");
+        assert_eq!(db.get_raw(42, 1).unwrap(), Some(vec![1, 2, 3]));
+    }
+
+    #[test]
+    fn guard_is_case_unsensitive() {
+        let tmpdir = tempfile::tempdir().unwrap();
+        init_app_dir(tmpdir.path()).unwrap();
+
+        let set_elem = || {
+            let db = open_app_dir(tmpdir.path(), false).unwrap();
+            db.put_raw(42, 1, vec![1, 2, 3].as_slice()).unwrap();
+        };
+        // lowercase 'y'
+        {
+            set_elem();
+            purge(tmpdir.path(), 42, None, None, Cursor::new("y")).expect("purge should succeed");
+            assert!(
+                open_app_dir(tmpdir.path(), true)
+                    .unwrap()
+                    .get(42, 1)
+                    .unwrap()
+                    .is_none()
+            );
+        }
+        // uppercase 'Y'
+        {
+            set_elem();
+            purge(tmpdir.path(), 42, None, None, Cursor::new("Y")).expect("purge should succeed");
+            assert!(
+                open_app_dir(tmpdir.path(), true)
+                    .unwrap()
+                    .get(42, 1)
+                    .unwrap()
+                    .is_none()
+            );
+        }
     }
 }
